+++ app/src/main/java/ru/androidschool/intensiv/data/Extensions.kt
@@ -1,0 +1,13 @@
+package ru.androidschool.intensiv.data
+
+import android.widget.ImageView
+
+class Extensions {
+
+    fun ImageView.loadImage(string: String) {
+
+
+
+    }
+
+}
